Total refactor of utils.py into separate utility modules and annotate all modules with an informative docstring
