+++ sequence_jacobian/__init__.py
@@ -1,11 +1,12 @@
 """Public-facing objects."""
 
-from . import asymptotic, determinacy, estimation, jacobian, nonlinear, utils
+from . import asymptotic, determinacy, estimation, jacobian, nonlinear, utilities
 from .models import rbc, krusell_smith, hank, two_asset
-from .blocks.simple_block import simple, apply_function
+from .blocks.simple_block import simple
 from .blocks.het_block import het
 from .blocks.helper_block import helper
 from .blocks.solved_block import solved
 from .blocks.combined_block import combine
+from .blocks.support.simple_displacement import apply_function
 from .steady_state import steady_state
 from .jacobian import get_G, get_H_U, get_impulse
