+++ sequence_jacobian/blocks/helper_block.py
@@ -1,6 +1,6 @@
 """HelperBlock class and @helper decorator to generate it"""
 
-from .. import utils
+from .. import utilities as utils
 
 
 def helper(f):
@@ -14,8 +14,8 @@ class HelperBlock:
 
     def __init__(self, f):
         self.f = f
-        self.input_list = utils.input_list(f)
-        self.output_list = utils.output_list(f)
+        self.input_list = utils.misc.input_list(f)
+        self.output_list = utils.misc.output_list(f)
         self.inputs = set(self.input_list)
         self.outputs = set(self.output_list)
