+++ sequence_jacobian/nonlinear.py
@@ -1,6 +1,8 @@
+"""Functions for solving for the non-linear transition dynamics provided a given shock path (e.g. solving MIT shocks)"""
+
 import numpy as np
 
-from . import utils
+from . import utilities as utils
 from . import jacobian as jac
 from .blocks import het_block as het
 
@@ -18,7 +20,7 @@ def td_solve(ss, block_list, unknowns, targets, H_U=None, H_U_factored=None, mon
     unknowns        : list, unknowns of SHADE DAG, the 'U' in H(U, Z)
     targets         : list, targets of SHADE DAG, the 'H' in H(U, Z)
     H_U             : [optional] array (nU*nU), Jacobian of targets with respect to unknowns
-    H_U_factored    : [optional] tuple, LU decomposition of H_U, save time by supplying this from utils.factor()
+    H_U_factored    : [optional] tuple, LU decomposition of H_U, save time by supplying this from utils.misc.factor()
     monotonic       : [optional] bool, flag indicating HetBlock policy for some k' is monotonic in state k
                                                                         (allows more efficient interpolation)
     returnindividual: [optional] bool, flag to return individual outcomes from HetBlock.td
@@ -52,10 +54,10 @@ def td_solve(ss, block_list, unknowns, targets, H_U=None, H_U_factored=None, mon
         if H_U is None:
             # not even H_U is supplied, get it (costly if there are HetBlocks)
             H_U = jac.get_H_U(block_list, unknowns, targets, T, ss, save=save, use_saved=use_saved)
-        H_U_factored = utils.factor(H_U)
+        H_U_factored = utils.misc.factor(H_U)
 
     # do a topological sort once to avoid some redundancy
-    sort = utils.block_sort(block_list, ignore_helpers=True)
+    sort = utils.graph.block_sort(block_list, ignore_helpers=True)
 
     # iterate until convergence
     for it in range(maxit):
@@ -70,7 +72,7 @@ def td_solve(ss, block_list, unknowns, targets, H_U=None, H_U_factored=None, mon
         else:
             # update guess U by -H_U^(-1) times errors H
             Hvec = jac.pack_vectors(results, targets, T)
-            Uvec -= utils.factored_solve(H_U_factored, Hvec)
+            Uvec -= utils.misc.factored_solve(H_U_factored, Hvec)
             Us = jac.unpack_vectors(Uvec, unknowns, T)
     else:
         raise ValueError(f'No convergence after {maxit} backward iterations!')
@@ -88,7 +90,7 @@ def td_map(ss, block_list, sort=None, monotonic=False, returnindividual=False, *
 
     # first get topological sort if none already provided
     if sort is None:
-        sort = utils.block_sort(block_list, ignore_helpers=True)
+        sort = utils.graph.block_sort(block_list, ignore_helpers=True)
 
     # TODO: Seems a bit weird that you pass variables ad hoc from the top-level
     #   e.g. calling nonlinear.td_solve() with rstar as a kwarg in one asset HANK.
