+++ sequence_jacobian/utilities/__init__.py
@@ -1,0 +1,3 @@
+"""Utilities relating to: interpolation, forward step/transition, grids and Markov chains, solvers, sorting, etc."""
+
+from . import differentiate, discretize, forward_step, graph, interpolate, misc, optimized_routines, solvers
